+++ src/main/java/LinkerBell/campus_market_spring/CampusMarketSpringApplication.java
@@ -4,9 +4,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.scheduling.annotation.EnableAsync;
 
 @ConfigurationPropertiesScan
 @EnableJpaAuditing
+@EnableAsync
 @SpringBootApplication
 public class CampusMarketSpringApplication {
