+++ src/main/java/LinkerBell/campus_market_spring/service/FcmService.java
@@ -3,14 +3,7 @@ package LinkerBell.campus_market_spring.service;
 import LinkerBell.campus_market_spring.domain.Item;
 import LinkerBell.campus_market_spring.domain.Keyword;
 import LinkerBell.campus_market_spring.dto.FcmMessageDto;
-import LinkerBell.campus_market_spring.global.error.ErrorCode;
-import LinkerBell.campus_market_spring.global.error.exception.CustomException;
 import LinkerBell.campus_market_spring.repository.UserFcmTokenRepository;
-import com.google.firebase.messaging.FirebaseMessaging;
-import com.google.firebase.messaging.FirebaseMessagingException;
-import com.google.firebase.messaging.Message;
-import com.google.firebase.messaging.MessagingErrorCode;
-import com.google.firebase.messaging.Notification;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class FcmService {
 
     private final UserFcmTokenRepository userFcmTokenRepository;
+    private final FcmNotificationService fcmNotificationService;
 
     @Value("${deeplink.keyword_url}")
     private String deeplinkKeywordUrl;
@@ -39,7 +33,7 @@ public class FcmService {
                     fcmToken,
                     savedItem);
 
-                sendNotification(sendingKeywordMessage);
+                fcmNotificationService.sendNotification(sendingKeywordMessage);
             }
 
         }
@@ -55,31 +49,4 @@ public class FcmService {
             .build();
     }
 
-    public void sendNotification(FcmMessageDto fcmMessageDto) {
-        Message.Builder messageBuilder = Message.builder()
-            .setToken(fcmMessageDto.getTargetToken())
-            .setNotification(Notification.builder()
-                .setTitle(fcmMessageDto.getTitle())
-                .setBody(fcmMessageDto.getBody())
-                .build());
-
-        if (fcmMessageDto.getDeeplinkUrl() != null) {
-            messageBuilder.putData("deeplink", fcmMessageDto.getDeeplinkUrl());
-        }
-
-        Message message = messageBuilder.build();
-        try {
-            String response = FirebaseMessaging.getInstance().send(message);
-            log.debug("Successfully sent message with deeplink: {}", response);
-        } catch (FirebaseMessagingException ex) {
-            if (ex.getMessagingErrorCode().equals(MessagingErrorCode.INVALID_ARGUMENT)) {
-                throw new CustomException(ErrorCode.FCM_INVALID_ARGUMENT);
-            } else if (ex.getMessagingErrorCode().equals(MessagingErrorCode.UNREGISTERED)) {
-                throw new CustomException(ErrorCode.FCM_UNREGISTERED);
-            } else {
-                throw new CustomException(ErrorCode.INTERNAL_SERVER_ERROR);
-            }
-        }
-
-    }
 }
